+++ src/utils/index.ts
@@ -1,3 +1,5 @@
 export { defaultSort, multiSort } from './sorting.utils';
 export { defaultFilter, applyFilters } from './filtering.utils';
+export { highlightMatch } from './highlight.utils';
+export { generateSampleData } from './generateSampleData';